requirements.txt
streamlit>=1.30.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
plotly>=5.0.0
wordcloud>=1.8.1
PyPDF2>=3.0.0
python-docx>=0.8.11
